extract multipagerow strategy to sepatated module
